+++ rust/src/config/mod.rs
@@ -1,0 +1,3 @@
+mod scaling_quality;
+
+pub use self::scaling_quality::ScalingQuality;
